ci: build and run tests, caching the build of itcoin-core keyed on its commit id
- gcc12 on ubuntu 22.04
- itcoin-core continues to be checked out on the side of itcoin-fbft. We are not
  using the automatic checkout feature of our cmake scripts
- the prolog tests and the C++ tests are folded together in "make test" target:
  it would be good to split those
- when bringing up the infrstructure for the tests, an adaptive retry strategy
  with timeouts and exponential backoff is used, in order to minimize the
  waiting time, but also be prepared to the occasional slowdown
- the retry script is taken from https://github.com/kadwanev/retry (license:
  Apache 2.0)
- the change also adds a badge to the README

The caching strategy is very coarse: as long as a successful build of
itcoin-core was performed in the past, it is reused, otherwise a full build is
done. Even with this simple stategy, when the itcoin-core build hits the cache,
a CI workflow takes 4-5 minutes instead of 35-40 minutes.

The list of itcoin-core libraries is already in thirdparty/CMakeLists.txt, so
we reuse it via a CMake trick (the CMake language does not offer any
straightforward way of doing it).

The usual solution for this kind of task involves writing a separate
somescript.cmake which does what we want (in this case simply print the
content of ITCOIN_CORE_LIBRARIES) and invoke it via cmake -P somescript.cmake.
We would then also need to extract the definition of ITCOIN_CORE_LIBRARIES in a
dedicated file in order to import it from both. This would be very cumbersome.

Let's instead abuse the target structure and just add a fake target that prints
what we want. In the CI yaml we then have to grep out the diagnostic message
"Built target print_itcoin_core_libraries", because CMake does not offer any
hooks for selectively disabling just a single diagnostic message (it's all or
nothing, and we want diagnostic messages in our builds).

One final issue is that the Azure runner has a "/bin/sh" whose "echo" command
does not understand the "-e" parameter, hence we have to force it to use bash
instead.